+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 - Fix iMac11,1 boot issues after PRAM reset
 - Fix DRM support on Nvidia-only configurations
   - Support optional setting between DRM and QuickSync support on iMacs13,x and iMac14,x
+- Add public beta support for Legacy GPU Acceleration
+  - Note ATI/AMD TeraScale 2 unsupported
 
 ## 0.1.0
 - Fix crash on iMacs with Metal GPUs